Add 'custom_formats' argument to generated code
Previously `custom_formats` was assumed to be global, which works fine
for evaluation, but results in errors when generating code.

The approach used to fix this problem was to add a second argument to
the `validate` functions when custom formats are provided.
For execution, the function is wrapped with a `partial` that hides the
extra argument (and therefore avoids changes in the API or documentation).