Preserving full path to a value when following $ref
This is just a "rebased" version of
https://github.com/horejsek/python-fastjsonschema/pull/92, that can be
compared and rebased into master.

I think this approach is really useful when debugging errors.
More than once I found myself trying to understand which property
exactly was failing. Preserving the complete path make is less confusing
and more straight forward to find the errors.

All the work was done by @ZbigniewRA in #92.