Fix `not: {}` schemas when given falsy instances.
These were improperly considering objects like None to be valid under
that schema.

The fix here simply adds {} as another special cased schema, since
presumably the `is True` block is meant as an optimization.

Note that there are of course many other things you can give to
`not` which disallow any possible instance, like `not: {"title": "foo"}`
but the fallback block should presumably catch them.

The block that was deleted here never looks correct to me (so I've
simply deleted it). Running the updated test suite would seem to
confirm, as this change takes the suite from:

    321 failed, 3268 passed, 12 deselected, 152 xfailed, 140 xpassed

to

    309 failed, 3280 passed, 12 deselected, 152 xfailed, 140 xpassed

i.e. a strict increase in compliance.